pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Skoldata_pipeline"
version = "0.1.0"
authors = [{name='Mattias', email='mattias.ivarsson@point.design'}]
description = "SKriv nått här"

requires-python = ">=3.8"

dependencies = [
    "typer",
    "duckdb",
    "dbt-duckdb",
    "dlt==1.18.2",
    "python-dotenv==1.2.1"
    # Add all your other dependencies here
]

[tool.setuptools.packages.find]
exclude = ["dbt", "data*", "dlt"]  # Exclude dbt, data, dlt folders



[project.scripts]
run-skola-pipeline = "ockero_data_pipeline.cli:app"  # This creates a command you can run from anywhere
